後端/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.7.1
    # via django
boto3==1.26.139
    # via -r requirements.in
botocore==1.29.139
    # via
    #   boto3
    #   s3transfer
django==4.2.1
    # via
    #   -r requirements.in
    #   django-storages
django-storages==1.13.2
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
python-dateutil==2.8.2
    # via botocore
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
typing-extensions==4.6.1
    # via asgiref
urllib3==1.26.16
    # via botocore
xlrd==0.9.4
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
